requirements.txt
tensorflow 
scikit-learn
logging 
pymongo 
numpy 
pandas 
pyyaml 
keras 
python-box 
dataclasses 
ensure 
mlflow 
dvc 
dagshub 
python-dotenv

